sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taco-agent"
version = "0.3.0"
description = "TACO — The A2A Construction Open-standard SDK"
readme = "../README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11",
    "a2a-sdk>=0.3.25",
]

[project.optional-dependencies]
server = ["a2a-sdk[http-server]>=0.3.25"]
client = ["httpx>=0.27"]
test = [
    "pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27",
    "a2a-sdk[http-server]>=0.3.25",
]
all = ["a2a-sdk[http-server]>=0.3.25", "httpx>=0.27"]
dev = [
    "a2a-sdk[http-server]>=0.3.25",
    "httpx>=0.27",
    "pytest>=8.0", "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "mypy>=1.14",
]

[project.scripts]
taco = "taco.cli:main"

[project.urls]
Homepage = "https://github.com/pelles-ai/taco"
Documentation = "https://taco-protocol.dev"
Repository = "https://github.com/pelles-ai/taco"

[tool.hatch.build]
packages = ["taco"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_untyped_defs = false
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "taco.agent_card"
disable_error_code = ["call-arg"]
